requirements.txt
# Deep Learning Framework
tensorflow==2.15.0

# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0

# Data Science & Visualization
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.2

# Jupyter Notebook (for running .ipynb files)
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1